backend/requirements.txt
# FastAPI and web server
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6

# Database
sqlalchemy==2.0.23
psycopg2-binary==2.9.9
alembic==1.12.1

# Authentication
passlib[argon2]==1.7.4
argon2-cffi==23.1.0
python-jose[cryptography]==3.3.0

# Configuration
pydantic==2.5.0
pydantic-settings==2.1.0
email-validator==2.1.0
python-dotenv==1.0.0

# OCR
pytesseract==0.3.10
Pillow==10.1.0

# NLP / ML (OPTIONAL - commented out by default)
# The OCR pipeline works with rule-based extraction and doesn't require these packages.
# Uncomment below if you want to use transformer models (not required for basic functionality):
# transformers>=4.35.0
# torch>=2.2.0
# Note: For CPU-only torch: pip install torch --index-url https://download.pytorch.org/whl/cpu

# Date utilities
python-dateutil==2.8.2

# Testing
pytest==7.4.3
pytest-asyncio==0.21.1
httpx==0.25.2

# Code quality
ruff==0.1.6
black==23.11.0

